Ajout de 2 tests unitaires pour le calcul RDM

+++ DLL_CTICM_RDM/CLASSES/Cls_DATA_RDM.vb
@@ -22,13 +22,13 @@
     End Structure
 
     Public Structure Struc_Output
-        Dim VZ(,) As Double                    'Effort tranchant dans l’élément i, aux deux extrémités (0 à NbNodes-2, 0 à 1)
-        Dim MYY(,) As Double                   'Moment fléchissant dans l’élément i, aux deux extrémités (0 à NbNodes-2, 0 à 1)
+        Dim VZ(,) As Decimal                    'Effort tranchant dans l’élément i, aux deux extrémités (0 à NbNodes-2, 0 à 1)
+        Dim MYY(,) As Decimal                   'Moment fléchissant dans l’élément i, aux deux extrémités (0 à NbNodes-2, 0 à 1)
 
-        Dim UZ() As Double                     'Déplacement vertical du nœud i (0 à NbNodes-1)
-        Dim ROTY() As Double                   'Rotation du nœud i (0 à NbNodes-1)
+        Dim UZ() As Decimal                     'Déplacement vertical du nœud i (0 à NbNodes-1)
+        Dim ROTY() As Decimal                   'Rotation du nœud i (0 à NbNodes-1)
 
-        Dim RZ() As Double                     'Réactions verticales aux nœuds support (0 à NbAppuis-1)
+        Dim RZ() As Decimal                     'Réactions verticales aux nœuds support (0 à NbAppuis-1)
     End Structure
 
 #End Region
